Geographic sorting for Twists (#70)
Twists are now ordered by proximity to the center of the viewport. Additionally, you can filter by Twists that are not your own.

This change involves a breaking migration. With the addition of the debug menu in v3.0, administrators can now save database state before upgrading. This will be required.

The database image has also changed from base postgres to postgis (and the version was incremented from 17 to 18).